--- packages/data/tests/unittest_utils/__init__.py
@@ -1,1 +1,0 @@
-../../../../test/unit/unittest_utils/__init__.py
