--- packages/data/tests/unittest_utils/utility.py
@@ -1,1 +1,0 @@
-../../../../test/unit/unittest_utils/utility.py
